Moved all commands to the connection rather than the pool
