+++ Sources/MySQL/Query/ConnectionPool+ForEach.swift
@@ -1,7 +1,7 @@
 import Async
 import Core
 
-extension ConnectionPool {
+extension Connection {
     /// Loops over all rows resulting from the query
     ///
     /// - parameter query: Fetches results using this query
@@ -10,28 +10,29 @@ extension ConnectionPool {
     /// - returns: A future that will be completed when all results have been processed by the handler
     @discardableResult
     internal func forEachRow(in query: Query, _ handler: @escaping ((Row) -> ())) -> Future<Void> {
-        return retain { connection, complete, fail in
-            // Set up a parser
-            let stream = RowStream(mysql41: connection.mysql41)
-            connection.receivePackets(into: stream.inputStream)
-            
-            stream.onClose = {
-                complete(())
-            }
-            
-            stream.errorStream = { error in
-                fail(error)
-            }
-            
-            stream.drain(handler)
-            
-            // Send the query
-            do {
-                try connection.write(query: query.string)
-            } catch {
-                fail(error)
-            }
+        let promise = Promise<Void>()
+        
+        let stream = RowStream(mysql41: self.mysql41)
+        self.receivePackets(into: stream.inputStream)
+        
+        stream.onClose = {
+            promise.complete(())
         }
+        
+        stream.errorStream = { error in
+            promise.fail(error)
+        }
+        
+        stream.drain(handler)
+        
+        // Send the query
+        do {
+            try self.write(query: query.string)
+        } catch {
+            promise.fail(error)
+        }
+        
+        return promise.future
     }
     
     /// Loops over all rows resulting from the query
@@ -43,27 +44,29 @@ extension ConnectionPool {
     /// - returns: A future that will be completed when all results have been processed by the handler
     @discardableResult
     public func forEach<D: Decodable>(_ type: D.Type, in query: Query, _ handler: @escaping ((D) -> ())) -> Future<Void> {
-        return retain { connection, complete, fail in
-            // Set up a parser
-            let resultBuilder = ModelStream<D>(mysql41: connection.mysql41)
-            connection.receivePackets(into: resultBuilder.inputStream)
-            
-            resultBuilder.onClose = {
-                complete(())
-            }
-            
-            resultBuilder.errorStream = { error in
-                fail(error)
-            }
-            
-            resultBuilder.drain(handler)
-            
-            // Send the query
-            do {
-                try connection.write(query: query.string)
-            } catch {
-                fail(error)
-            }
+        let promise = Promise<Void>()
+        
+        // Set up a parser
+        let resultBuilder = ModelStream<D>(mysql41: self.mysql41)
+        self.receivePackets(into: resultBuilder.inputStream)
+        
+        resultBuilder.onClose = {
+            promise.complete(())
         }
+        
+        resultBuilder.errorStream = { error in
+            promise.fail(error)
+        }
+        
+        resultBuilder.drain(handler)
+        
+        // Send the query
+        do {
+            try self.write(query: query.string)
+        } catch {
+            promise.fail(error)
+        }
+        
+        return promise.future
     }
 }
